ads, characters, game, gold, howToPlay, mainMenu
Each .lua file corresponds to its respective scene in the Current Crosser app. For example, game.lua is run when the player is playing the main game in the app, and howToPlay.lua is run when the player is in the "How to Play" scene. main.lua is excluded here because it contains personal info necessary for ads and in-app purchases. Current Crosser was created using the Lua coding language and Corona SDK.